Creating Welcome and Finish
Creating a blank welcome screen and a finish screen.
Created a Game class that will host the Level, Welcome and Finish
methods. Such methods are connected between each other by passing
themselves as a parameter of the other methods. eg. Level method
will pass the finish method as an argument that can be then triggered
in the Level Class.

Created the win and die conditions of the players. (could
they be simplified ? )