Points mechanic for players
Created a new file dedicated to the points and points bar
for the players. It checks if the player is colliding with
the Edibles class, and if so, it takes the assigned value of
the edible (1 or -1). This is reflected on the points counter,
next step is connecting the points bar to the same behaviour.