Adding Winner Screens and Game Over
Added a winner screen for each player.
Game over window will display when both characters have fallen
from the level.
The game can be played again, or can be closed, and point counter
will be restarted.
I have also added sounds to the Second player when it
collides with the crypto.